requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
langgraph==0.2.28
openai>=1.40.0
pymupdf==1.24.10
python-docx==1.1.2
chromadb==0.5.5
requests==2.32.3
python-dotenv==1.0.1